% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\docType{package}
\name{GHtools-package}
\alias{GHtools-package}
\title{Manage Project Files and Repositories on GitHub using R.}
\description{
GitHub is one of the best-known Web Services for software development and 
version control using Git. It supports access control, bug tracking, 
software feature requests, task management and much more. 
It is very easy to interact with GitHub from a Bash/Unix terminal 
using git. Here, we provide a collection of functions to access 
GitHub from an R environment and perform some simple basic operations.
}
\references{
\enumerate{
  \item{\strong{GitHub Website}: \url{https://github.com/} }
  \item{\strong{GitHub Wiki Page}: \url{https://en.wikipedia.org/wiki/GitHub} }
  \item{\strong{Get started with Git}: \url{https://www.w3schools.com/git/git_getstarted.asp?remote=github} }
 }
}
