% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_checkout_branch}
\alias{git_checkout_branch}
\title{Checkout into a GitHub Repo Branch.}
\usage{
git_checkout_branch(
  git_folder = ".",
  branch = "master",
  force = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{git_folder}{path to a local folder where to a GitHub repo
has been cloned/installed. Defaults to '.' (local directory).}

\item{branch}{string corresponding to the branch to checkout into.}

\item{force}{logical (defaults to TRUE). If branch does not exist, create
one with this name.}

\item{verbose}{logical (defaults to TRUE). Shall information be printed 
to console.}
}
\description{
Checkout into a specific branch of a local copy of a GitHub repository.
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
