% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_clone_repo}
\alias{git_clone_repo}
\title{Clone a GitHub Repository.}
\usage{
git_clone_repo(repo, user = NULL, token = NULL, branch = NULL, dest = NULL)
}
\arguments{
\item{repo}{string, GitHub repository to clone (e.g., 'dami82/mutSignatures')}

\item{user}{string (optional). Identifier of the Github user 
attempting the repo cloning. Defualts to NULL (no user ID).}

\item{token}{string (optional). Authentication Token for the user
attempting the repo cloning. Defualts to NULL (no token).}

\item{branch}{string (optional). Name of the single branch to download. 
Defaults to NULL (if so, all branches are retrieved).}

\item{dest}{path to a local folder where to save the downloaded files. 
Defaults to '.' (install in a subfolder in the local directory).}
}
\value{
code returned by the 'git clone' command
}
\description{
Download a local copy of a GitHub repository.
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
