% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_commit_and_push}
\alias{git_commit_and_push}
\title{Commit and Push to GitHub.}
\usage{
git_commit_and_push(
  git_folder = ".",
  branch = "master",
  commit_message = NULL,
  user = NULL,
  token = NULL,
  files = NULL,
  force = TRUE
)
}
\arguments{
\item{git_folder}{path to a local folder where to a GitHub repo
has been cloned/installed. Defaults to '.' (local directory).}

\item{branch}{string corresponding to the branch to commit and push
to GitHub (defaults to 'master').}

\item{commit_message}{string, commit message describind the file changes.}

\item{user}{string (optional). Identifier of the Github user 
attempting the repo cloning. Defualts to NULL (no user ID).}

\item{token}{string (optional). Authentication Token for the user
attempting the repo cloning. Defualts to NULL (no token).}

\item{files}{character vector of filenames (paths to files) to be added to
the commit. Defaults to NULL (all files in the `git_folder` will be added).}

\item{force}{logical (defaults to TRUE). The -f tag is attached
to the 'git push' command.}
}
\value{
NULL.
}
\description{
Add and Commit File Changes and then Push all modifications
remotely to the GitHub repository.
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
