% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_list_all_branches}
\alias{git_list_all_branches}
\title{List All Branches of a GitHub Repo.}
\usage{
git_list_all_branches(git_folder = ".")
}
\arguments{
\item{git_folder}{path to a local folder where to a GitHub repo
has been cloned/installed. Defaults to '.' (local directory).}
}
\value{
list of all repo branches. The list includes up to 3 elements (`starred`, `local`, and `remote`); 
each element is a character vector including the branch names.
}
\description{
List all branches of a GitHub repository, including local and 
remote branches. The currently active (starred) branch is also 
identified and returned.
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
