% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_update_repo_from_folder}
\alias{git_update_repo_from_folder}
\title{Update GitHub Repo from a Local Folder.}
\usage{
git_update_repo_from_folder(
  repo,
  local_folder,
  gh_user_id,
  token = NULL,
  gh_user_fullname,
  gh_user_email,
  branch = NULL,
  commit_message = NULL,
  temporary_folder = "tmp_ghdir",
  force = FALSE
)
}
\arguments{
\item{repo}{string, GitHub repository to clone (e.g., 'dami82/mutSignatures')}

\item{local_folder}{path to a local folder that include all source files (and
folders) to be uploaded to GitHub.}

\item{gh_user_id}{string (optional). Identifier of the Github user 
attempting the repo update operation.}

\item{token}{string (optional). Authentication Token for the user
attempting the repo update operation. Defualts to NULL (no token).}

\item{gh_user_fullname}{string (optional). Full name of the Github user 
attempting the repo update operation.}

\item{gh_user_email}{string (optional). Email address of the Github user 
attempting the repo update operation.}

\item{branch}{string (optional). Name of the single branch to download. 
Defaults to NULL (if so, all branches are retrieved).}

\item{commit_message}{string, commit message describing the file changes.}

\item{temporary_folder}{name of a temporary folder that is used to
save temporary copies of all files available on GitHub. An error is
raised if this folder already exists on the local machine, unless the 
`force` argument is set to `TRUE`.}

\item{force}{logical, shall files and folders be overwritten if needed.}
}
\value{
NULL.
}
\description{
Update a GitHub repo by uploading all files included in a local folder.
The files can be written in a new or existing branch. Hidden files and
folders in `local_folder`  will be ignored (this includes sub-folder 
such as `.git/`). Note that the `.git/` hidden subfolder is 
retrieved from GitHub during the operation).
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
