% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_v02.R
\name{git_which_active_branch}
\alias{git_which_active_branch}
\title{Check Active GitHub Branch.}
\usage{
git_which_active_branch(git_folder = ".")
}
\arguments{
\item{git_folder}{path to a local folder where to a GitHub repo
has been cloned/installed. Defaults to '.' (local directory).}
}
\value{
string corresponding to the active branch.
}
\description{
Check which is the Active Branch of a 
local installation of a GitHub repository.
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
